+++ resources/lang/en/errors.php
@@ -112,6 +112,12 @@ return [
         'code' => '',
     ],
 
+    'resource_cannot_be_deleted' => [
+        'title' => 'Not Deletable',
+        'detail' => 'The resource cannot be deleted.',
+        'code' => '',
+    ],
+
     'query_invalid' => [
         'title' => 'Invalid Query Parameter',
         'detail' => 'The request query parameters are invalid.',