+++ src/Http/Requests/DeleteResource.php
@@ -60,6 +60,11 @@ class DeleteResource extends ValidatedRequest
         $this->passes(
             $validators->modifyQuery($this->query())
         );
+
+        /** Validate that the delete is allowed. */
+        if ($validator = $validators->delete($this->getRecord())) {
+            $this->passes($validator);
+        }
     }
 
 }